Verify graphs so they are unique
Not we check in the parent population and in the generation that we are
building that we don't have any duplicates.  Turns out that this is very
often the case, unfortunately.

The functions could (and probably should) be refactored to check the
uniqueness of the graph in one bigger call.  That would also have the
benefit of incremetally merging the current population and the next
one.  Not sure if there's acutally time for that, but I would be
surprised if there was.